Rewrote commands system and command _FIRST_POKE_SELECT_PROC
