+++ src/mod/features/commands/utils/utils.h
@@ -1,0 +1,28 @@
+#pragma once
+
+#include "externals/il2cpp-api.h"
+
+#include "externals/EvData/Aregment.h"
+#include "externals/Pml/PokePara/PokemonParam.h"
+
+// Converts the IEEE-754 formatted int32 into its proper float value.
+float ConvertToFloat(int32_t value);
+
+// Returns either the value at the given work variable or the direct int value depending on the given argument's type.
+// Returns 0 for any other argument type.
+int32_t GetWorkOrIntValue(EvData::Aregment::Object arg);
+
+// Returns either the value at the given work variable or the direct float value depending on the given argument's type.
+// Returns 0 for any other argument type.
+float GetWorkOrFloatValue(EvData::Aregment::Object arg);
+
+// If the given argument is a work variable, set it to the given value.
+// Does nothing otherwise.
+void SetWorkToValue(EvData::Aregment::Object arg, int32_t value);
+
+// Checks if the given PokemonParam is null or an egg.
+bool IsNullOrEgg(Pml::PokePara::PokemonParam::Object * param);
+
+// Tries to add a new Pokémon to the party.
+// Returns false if it fails.
+bool AddPokemonToParty(int32_t monsno, int32_t formno, uint32_t level, uint8_t maxedIVs, uint16_t itemno);
